refactor(ProductList): replace product list with component
