+++ customTypes/index.ts
@@ -1,4 +1,10 @@
+import { Product } from '.prisma/client';
+
 export interface ResponseType {
   ok: boolean;
   [key: string]: any;
+}
+
+export interface ProductWithCount extends Product {
+  _count: any;
 }
